feat: add benchmark-vm flake application
Create `scripts/benchmark-vm/{benchmark-vm.sh,default.nix}` and add a
corresponding entrypoint in `flake.nix`.

This adds two tests for measuring the performance of the Virby VM. One measures
the time it takes to boot the VM from a cold start, and the other measures the
time it takes to build a derivation (default: `nixpkgs#hello`), and exports the
results to a file.