fix: force ssh to use IPv4 and only the generated SSH key
The launchd daemon only listen to IPv4 but ssh may resolve localhost as
::1 preventing the onDemand boot to work.

This commit explicitly tells ssh to use IPv4 to resolve virby-vm

It also instructs ssh to only attempt a login with the specified
IdentityFile. This is important to avoid the "too many login attempts"
error in a system with many ssh identities in the ssh agent.